Made the main ESP code with char* instead of string, may solve the freeze out issues. Also when getting system id from api it saves a default config to the system, can be edited from app anytime. Need to set up fetching config data from server on power up, also how to auto update it on change.

+++ ATC_ESP_program/lib/libAQWiFiConfig/AQWiFiConfig.h
@@ -1,9 +1,8 @@
 #ifndef AQESPWifiConfig_h
 #define AQESPWifiConfig_h
-#include "Arduino.h"
-#include "DebugUtils.h"
 #include "ESP8266WebServer.h"
 #include "MemoryHandler.h"
+#include <Arduino.h>
 
 //////////////////////////////////////////////////////////////////
 // This class is managing the WiFi connection and the credentials for the network
@@ -15,7 +14,7 @@ private:
     static bool isConfigDoneFlag;
     String WIFI_SSID;
     String WIFI_PASS;
-    uint16_t systemID;
+    uint8_t systemID;
     static MemoryHandler* memHandler;
     static ESP8266WebServer* configServer;
     static void createSiteForWiFiLogin();
@@ -27,7 +26,7 @@ public:
     ~AQWiFiConfig() {};
     String getSSID();
     String getPassword();
-    uint16_t getSystemID();
+    uint8_t getSystemID();
     void saveSystemID(const uint16_t& id);
     void forgetNetwork();
     void initializeNetwork();
